This commit fixes an error, at least I think that, in the Python interface of Akantu.
The problem is that several parts needs the `ElementTypeMapArray` type, such as the mesh, the FEEngine and the material, probably even more, but grep is not able to find them (because of the use of `auto`).
This type is indirectly registered by the `register_material()` function in the `py_material.cc` file.
However that function is only run if the solid mechanics part is enabled.
Since this is almost always the case, it was probably not an issue till now.
However since I have deactivated it, for my development of the structural mechanics part, I run into that issue, there is still one, but I have not found it, I am not even sure if _this_ is an error.

The fix is that the `ElementTypeMapArray` type is now registered by the `register_mesh()` function, which is the first occurrence of that type (as far as I can tell) and is always built.

An important side note is, that the `InternalField` is sill registered by the material.