Updated the Interface of the mesh to allow the creation of the `extra_normal` data in the mesh.
The fix is not very pretty, however it allows to create and access the `extra_normal` data from Python.
There are three functions, one to check for its existence, one for creating the property, if it does not exists and one for accessing it.
I am not sure if this is the best way to do it, however it seams to work.

I also created a new test for verifying it.