Updated the text description of the `getMaterial()` functions of the structural mechanic model.
The new text makes it clear that the materials are returned as references and not as copies as before.
There is one aspect that I do not like however.
The `getMaterial()` function operating on names, i.e. strings, uses the argument `i`, which is consistent with the other version, however I would suggest to rename it to `name` which is more logical.